refactor: InputHandler의 역할 분
